Update README.md
Morning team, just updated to what was discussed in class. I don't have notes on the solution brainstorm and we were not able to finish the class discussion yesterday on the elevator pitch. I've placed my suggestion in the elevator pitch so we can catch up to current. Let me know what you think?